my_code/requirements.txt
pandas
addict
fire
numpy
tqdm
scipy
scikit-learn
pyyaml
numba
matplotlib
h5py
openpyxl
lightgbm
optuna